Worked on adding parallel merge

+++ src/main.h
@@ -8,11 +8,11 @@ typedef uint32_t vec_t;
 
 struct AlgoArgs {
     vec_t* A;
-    uint64_t A_length;
+    uint32_t A_length;
     vec_t* B;
-    uint64_t B_length;
+    uint32_t B_length;
     vec_t* C;
-    uint64_t C_length;
+    uint32_t C_length;
     vec_t* CUnsorted;
     uint32_t threadNum;
     uint32_t numThreads;
@@ -22,9 +22,9 @@ struct AlgoArgs {
 };
 
 void MergePathSplitter(
-    vec_t * A, uint64_t A_length,
-    vec_t * B, uint64_t B_length,
-    vec_t * C, uint64_t C_length,
+    vec_t * A, uint32_t A_length,
+    vec_t * B, uint32_t B_length,
+    vec_t * C, uint32_t C_length,
     uint32_t threads, uint32_t* ASplitters, uint32_t* BSplitters);
 
 int hostBasicCompare(const void * a, const void * b);
